move connection logging to a separate file, absorb logutils package
No functional change expected.
