--- internal/logutils/logutils_suite_test.go
@@ -1,13 +1,0 @@
-package logutils
-
-import (
-	"testing"
-
-	. "github.com/onsi/ginkgo/v2"
-	. "github.com/onsi/gomega"
-)
-
-func TestLogutils(t *testing.T) {
-	RegisterFailHandler(Fail)
-	RunSpecs(t, "Logutils Suite")
-}
